[#398] unite ShareButton and Popup components to ShareButtonWithPopover, some fix
