+++ apps/blog/src/components/ShareButtonWithPopover/index.ts
@@ -1,0 +1,1 @@
+export { ShareButtonWithPopover } from './ShareButtonWithPopover'
